requirements.txt
langchain>=0.2.11
langgraph>=0.2.14
langchain-openai>=0.1.20
openai
langchain-community>=0.2.11
pydantic>=2.8
python-dotenv>=1.0.1
faiss-cpu>=1.8.0
google-search-results
keybert
streamlit
langchain-anthropic
scikit-learn
sentence-transformers
newspaper3k
numpy>=1.24.0
tiktoken>=0.7.0
sentence-transformers>=2.7.0
scikit-learn>=1.3.0
pdfminer.six>=20231228
requests>=2.31.0
beautifulsoup4>=4.12.0
serpapi
playwright
beautifulsoup4
readability-lxml
sentence-transformers
scikit-learn
dateparser
validators
pytest
